a
git-svn-id: http://aproject.unfuddle.com/svn/aproject_androidvj@10 13ae8a00-e42d-4a4a-b426-7c3bbda1a15a

+++ aVideogame/src/avideogame/domain/HotSpot.java
@@ -1,10 +1,10 @@
 package avideogame.domain;
 
 public class HotSpot {
-	private int x;
-	private int y;
-	private int width;
-	private int height;
+	private double x;
+	private double y;
+	private double width;
+	private double height;
 	
 	
 	/**
@@ -13,31 +13,31 @@ public class HotSpot {
 	 * @param py position y to check
 	 * @return true if position is inside hotspot, false otherwise
 	 */
-	public boolean isInside(int px, int py){
+	public boolean isInside(double px, double py){
 		if(px < x+width && px > x && py < y+height && py > y) return true;
 		else return false;
 	}
 	
 	//GETTERS & SETTERS
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
 	public void setWidth(int width) {
 		this.width = width;
 	}
-	public int getHeight() {
+	public double getHeight() {
 		return height;
 	}
 	public void setHeight(int height) {
 		this.height = height;
 	}
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	public void setX(int x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	public void setY(int y) {
